Refactored normal mode operations into their own functions
Took a page off of neovim's operator handling in this.
With this we temporarily lost functonal key support (arrows etc.)
for insert mode. OTOH all the operators are now neatly
compartmentalized.
Some are still a bit hacky -- we need to pass more complete information
about the state of the editor for some operators to work correctly (e.g.
'l' 'go to bottom line' needs to know whether wrap is on and what
the screen dimensions are).