Escape insert mode: move cursor back one char (like vim)
In insert mode the cursor is between the characters, while in normal
mode it is on a character. Since it is appropriate in insert mode
to see the character before the cursor as the one being edited (esp. at
the end of the line), we want to move the cursor onto the current
character when we return to normal mode, for consistency's sake.